refactor(cmd/pilot): extract adapter wrappers to adapters.go (GH-1213)
Move adapter wrapper types and helper functions from main.go to a dedicated
adapters.go file to improve code organization before further refactoring.

Extracted:
- telegramBriefAdapter type and methods
- telegramApprovalAdapter type and methods
- convertKeyboardToTelegram helper
- slackApprovalClientAdapter type and methods
- wireProjectAccessChecker function
- getAlertsConfig function
- qualityCheckerWrapper type and methods
- resolveOwnerRepo function
